# Conflicts:
#	LICENSE
#	README.md
#	array_test.go
#	bench_array_test.go
#	bench_example_test.go
#	bench_expand_test.go
#	bench_filter_test.go
#	bench_iteration_test.go
#	bench_property_test.go
#	bench_query_test.go
#	bench_traversal_test.go
#	doc.go
#	doc/tips.md
#	example_test.go
#	expand.go
#	expand_test.go
#	filter.go
#	filter_test.go
#	iteration.go
#	manipulation.go
#	manipulation_test.go
#	misc/git/pre-commit
#	property.go
#	property_test.go
#	query.go
#	query_test.go
#	traversal.go
#	traversal_test.go
#	type.go
#	type_test.go
#	utilities.go

+++ src/jk/jkservice/jkstatus/instance.go
@@ -26,7 +26,7 @@ func NewRemoteInstance(conn net.Conn, inter int) (*RemoteInstance, error) {
 
 func (ri *RemoteInstance) Update() {
 	now := time.Now().Unix()
-	if now-ri.Last > ri.Interval {
+	if now-ri.Last > int64(ri.Interval) {
 		ri.Online = false
 	}
 }
